requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bitarray==3.8.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
coverage==7.13.4
cryptography==46.0.5
cytoolz==1.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
kalshi-python==2.1.4
lazy_imports==1.2.0
librt==0.8.1
mypy==1.19.1
mypy_extensions==1.1.0
packaging==26.0
parsimonious==0.10.0
pathspec==1.0.4
pluggy==1.6.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
regex==2026.2.28
requests==2.32.5
ruff==0.15.6
six==1.17.0
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
